break up server schema into multiple files
