+++ v2/server/schema/venueInputs.js
@@ -1,0 +1,11 @@
+import { gql } from 'apollo-server'
+
+export default gql`
+  input CityInput {
+    id: Int!
+  }
+
+  input VenueTypeInput {
+    id: Int!
+  }
+`
